+++ tests/Variables.Tests.ps1
@@ -50,25 +50,31 @@ Describe 'Variables' {
             $orgVariableName = "$variableName`_ORG"
             $environmentName = "$testName-$os-$TokenType-$id"
 
-            switch ($OwnerType) {
-                'user' {
-                    $repo = Set-GitHubRepository -Name $repoName -AddReadme -License 'mit' -Gitignore 'VisualStudio'
-                }
-                'organization' {
-                    Get-GitHubVariable -Owner $Owner | Where-Object { $_.Name -like "$variableName*" } | Remove-GitHubVariable -Confirm:$false
-                    $repo = Set-GitHubRepository -Organization $Owner -Name $repoName -AddReadme -License 'mit' -Gitignore 'VisualStudio'
-                    $repo2 = Set-GitHubRepository -Organization $Owner -Name "$repoName-2"
-                    $repo3 = Set-GitHubRepository -Organization $Owner -Name "$repoName-3"
-                    LogGroup "Org variable - [$orgVariableName]" {
-                        $params = @{
-                            Owner                = $owner
-                            Name                 = $orgVariableName
-                            Value                = 'organization'
-                            Visibility           = 'selected'
-                            SelectedRepositories = $repo.id
+            if ($OwnerType -in ('repository', 'enterprise')) {
+                $repo = $null
+                $repo2 = $null
+                $repo3 = $null
+            } else {
+                switch ($OwnerType) {
+                    'user' {
+                        $repo = Set-GitHubRepository -Name $repoName -AddReadme -License 'mit' -Gitignore 'VisualStudio'
+                    }
+                    'organization' {
+                        Get-GitHubVariable -Owner $Owner | Where-Object { $_.Name -like "$variableName*" } | Remove-GitHubVariable -Confirm:$false
+                        $repo = Set-GitHubRepository -Organization $Owner -Name $repoName -AddReadme -License 'mit' -Gitignore 'VisualStudio'
+                        $repo2 = Set-GitHubRepository -Organization $Owner -Name "$repoName-2"
+                        $repo3 = Set-GitHubRepository -Organization $Owner -Name "$repoName-3"
+                        LogGroup "Org variable - [$orgVariableName]" {
+                            $params = @{
+                                Owner                = $owner
+                                Name                 = $orgVariableName
+                                Value                = 'organization'
+                                Visibility           = 'selected'
+                                SelectedRepositories = $repo.id
+                            }
+                            $orgVariable = Set-GitHubVariable @params -Debug
+                            Write-Host ($orgVariable | Select-Object * | Out-String)
                         }
-                        $orgVariable = Set-GitHubVariable @params -Debug
-                        Write-Host ($orgVariable | Select-Object * | Out-String)
                     }
                 }
             }
